# internal/schema/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAgentPlatform API
  version: 1.1.0
  description: |
    REST API for OpenAgentPlatform (OAP) — an open-source, agent-first
    Remote Monitoring and Management (RMM) platform.

    ## Authentication
    All `/api/v1/*` endpoints require a bearer token (OIDC ID token or
    session JWT). Auth endpoints under `/auth/*` use OIDC Authorization
    Code Flow with PKCE.

    ## Versioning
    The API is versioned via URL path (`/api/v1`). Breaking changes
    increment the major version; additive changes are backward compatible.

    ## Rate Limiting
    Default rate limit is 1000 requests/min per token. Burst to 2000.
  contact:
    name: OpenAgentPlatform
    url: https://openagentplatform.io
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.openagentplatform.io
    description: Production
  - url: https://staging-api.openagentplatform.io
    description: Staging
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/agents:
    get:
      tags: [Agents]
      summary: List agents
      description: Returns a paginated list of registered agents.
      operationId: listAgents
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: site_id
          in: query
          description: Filter by site ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by agent status
          schema:
            type: string
            enum: [online, offline, degraded, pending]
      responses:
        "200":
          description: Paginated agent list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAgent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/alerts:
    get:
      tags: [Alerts]
      summary: List alerts
      operationId: listAlerts
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [firing, acknowledged, resolved, snoozed]
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warning, critical]
        - name: agent_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated alert list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAlert"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/alerts/{id}/acknowledge:
    post:
      tags: [Alerts]
      summary: Acknowledge an alert
      operationId: acknowledgeAlert
      parameters:
        - $ref: "#/components/parameters/AlertIdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  description: Optional acknowledgment note
      responses:
        "200":
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/alerts/{id}/resolve:
    post:
      tags: [Alerts]
      summary: Resolve an alert
      operationId: resolveAlert
      parameters:
        - $ref: "#/components/parameters/AlertIdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                resolution:
                  type: string
                  description: Resolution reason or note
      responses:
        "200":
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/alerts/{id}/snooze:
    post:
      tags: [Alerts]
      summary: Snooze an alert
      operationId: snoozeAlert
      parameters:
        - $ref: "#/components/parameters/AlertIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [until]
              properties:
                until:
                  type: string
                  format: date-time
                  description: When to resume notifications
                reason:
                  type: string
      responses:
        "200":
          description: Alert snoozed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/alert-rules:
    get:
      tags: [AlertRules]
      summary: List alert rules
      operationId: listAlertRules
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: enabled
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated alert rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAlertRule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [AlertRules]
      summary: Create alert rule
      operationId: createAlertRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertRuleCreate"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/alert-rules/{id}:
    put:
      tags: [AlertRules]
      summary: Update alert rule
      operationId: updateAlertRule
      parameters:
        - $ref: "#/components/parameters/AlertRuleIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertRuleUpdate"
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # =========================================================
  # Policies (Phase 1)
  # =========================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC ID token or session JWT
  parameters:
...
    PageParam:
      name: page
      in: query
      description: 1-based page index
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Items per page (max 200)
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    AlertIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AlertRuleIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized — missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden — token lacks required role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: invalid_request
        message:
          type: string
          description: Human-readable error message
        request_id:
          type: string
          description: Correlation ID for log lookup
    PageMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
...
    Agent:
      type: object
      required: [id, hostname, status, org_id, created_at]
      properties:
        id:
          type: string
          format: uuid
        hostname:
          type: string
        os:
          type: string
          enum: [linux, darwin, windows]
        os_version:
          type: string
        arch:
          type: string
          enum: [amd64, arm64, 386, arm]
        status:
          type: string
          enum: [online, offline, degraded, pending]
        last_seen_at:
          type: string
          format: date-time
        agent_version:
          type: string
        site_id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        labels:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Alert:
      type: object
      required: [id, status, severity, rule_id, created_at]
      properties:
        id:
          type: string
          format: uuid
        rule_id:
          type: string
          format: uuid
        agent_id:
          type: string
          format: uuid
        check_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [firing, acknowledged, resolved, snoozed]
        severity:
          type: string
          enum: [info, warning, critical]
        title:
          type: string
        description:
          type: string
        acknowledged_by:
          type: string
        acknowledged_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
        snoozed_until:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        labels:
          type: object
          additionalProperties:
            type: string
    AlertRule:
      type: object
      required: [id, name, expression, enabled, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        expression:
          type: string
          description: CEL or expression evaluated against check results
          example: check.status == "fail" && check.severity == "critical"
        severity:
          type: string
          enum: [info, warning, critical]
        enabled:
          type: boolean
        notification_channels:
          type: array
          items:
            type: string
            enum: [email, slack, webhook, pagerduty]
        cooldown_seconds:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AlertRuleCreate:
      type: object
      required: [name, expression]
      properties:
        name:
          type: string
        description:
          type: string
        expression:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        enabled:
          type: boolean
          default: true
        notification_channels:
          type: array
          items:
            type: string
            enum: [email, slack, webhook, pagerduty]
        cooldown_seconds:
          type: integer
    AlertRuleUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        expression:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        enabled:
          type: boolean
        notification_channels:
          type: array
          items:
            type: string
        cooldown_seconds:
          type: integer
...
    PaginatedAgent:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        page:
          $ref: "#/components/schemas/PageMeta"
...
    PaginatedAlert:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        page:
          $ref: "#/components/schemas/PageMeta"
    PaginatedAlertRule:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AlertRule"
        page:
          $ref: "#/components/schemas/PageMeta"
...